requirements.txt
torch
pandas
numpy
tqdm
pyyaml
matplotlib
wget
scikit-learn
